list item create loop
