+++ app/src/main/java/com/mjkj/listit/Activity/ScrollListsActivity.kt
@@ -4,21 +4,27 @@ import android.annotation.SuppressLint
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.mjkj.listit.Composable.ListAppBar
-import com.mjkj.listit.Composable.ListItem
 
 class ScrollListsActivity : ComponentActivity() {
     @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
@@ -37,7 +43,6 @@ class ScrollListsActivity : ComponentActivity() {
                         )
                     }
                 ) {
-
                     Column(
                         modifier = Modifier
                             .fillMaxSize()
@@ -45,16 +50,42 @@ class ScrollListsActivity : ComponentActivity() {
                         verticalArrangement = Arrangement.Top,
                     ) {
                         Spacer(modifier = Modifier.height(100.dp))
-                        val items = List(10) { index -> "Item ${index + 1}" }
-
                         LazyColumn {
-                            items(items) { item ->
-                                ListItem(title = item, description = "To jest opis dla $item")
+                            for (i in 1..3) {
+                                item {
+                                    ListItem(title = "Example ${i}", description = "Example${i}")
+                                }
                             }
                         }
                     }
                 }
             }
+        }
+    }
+}
+
+@Composable
+fun ListItem(title: String, description: String) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(10.dp)
+            .background(Color.LightGray, shape = RoundedCornerShape(8.dp))
+    ) {
+        Column(
+            modifier = Modifier.padding(horizontal = 10.dp, vertical = 8.dp)
+        ) {
+            Text(
+                text = title,
+                color = Color.Black,
+                fontSize = 25.sp
+            )
+            Spacer(modifier = Modifier.height(8.dp))
+            Text(
+                text = description,
+                color = Color.Black,
+                fontSize = 16.sp
+            )
         }
     }
 }
